Fix MultiVisitor to not call None visit functions (#20)
Fixes #19: The MultiVisitor was not skipping any None visit functions returned by its child Visitors.

Note that we always want all the visitors to visit all the nodes in the hierarchy, even if they are to be skipped. Each visitor at a node should decide what to do if the node is skipped. (Rationale: we want to be able to report on the skipped nodes to, even if no tests executed there.)

I'll add tests to MultiVisitor separately (#21)