Implement assert_contains_any and related checks (#61)
* Implement assert_contains_any and related checks

We now allow checking that all elements are or are not included in the
last output, and that some elements are or are not included in the
last output

In `TestCase`;
- rename `_check_contains` to `_check_several`
- have `_check_contains` take `any` or `all`
- rename `assert_*contains` methods to clarify semantics (and similar for `expect` variants)
- expand documentation of these methods
- implement {expect,assert}_contains_any
- implement {expect,assert}_not_contains_some

Add tests for the above

* For the `contains` functions, make `message` be a kwarg

This affects only the code invocation at the moment.

In the future, we can change the yaml invocation so that message need
not be the first element in the list. We could also do this for case sensitivity.

* Simplify string inclusion check

When registering the string inclusion/exclusion functions, generate
these functions automatically based on the various condition
variants. This reduces code duplication and clarifies the semantics.

* Update documentation for including/excluding any/all strings