Support implicit tags: @manifest_source, @manifest_dir
This adds support for implicit tags, which are tags not defined
explicitly in the manifest docs but which are applied automatically to
all elements in a given manifest doc.

- In particular, all YAML documents in the same YAML stream get the
  same implicit tags, but documents read from different streams will
  in general have different implicit tags.

- This means that a manifest held in memory could have different sets
  of elements, each set with different implicit tags, since in the
  general case the manifest in memory could come from multiple YAML
  streams (each containing multiple YAML documents).

This commit implements two such implicit tags, `@manifest_source` (the full
path to the source manifest file , including the filename itself), and
`@manifest_dir` (the directory of the manifest file).

- When manifests are passed from a string (as happens in testing this
  code or maybe in the future when we read from stdin),
  `@manifest_dir` is empty and `@manifest_source` is whatever
  descriptive string is provided by caller.