Make setup failures preempt test case (#90)
Each test case is preceded by a setup run specified at the suite level. With this change, if the setup fails, the test case body is not run and we jump directly to the case teardown (also specified at the suite level). Note that if the the suite contains multiple test cases, this process will happen for each test case where the setup fails.