Make basepath configurable, defaulting to '.' (#93)
We still use `basepath` as before, but it now defaults to . so that the path to every sample is relative (as calculated from the cwd in which `gen-manifest` was run) by default.

If `--basepath=xxx` is specified, we still find samples relative to the cwd, but in the emitted manifest we use the provided value of `basepath`. This allows us to rebase the manifest to a different directory in anticipation of the samples being copied there.