Allow unescaped `@` in test.yaml
Previously, any uses of `@` in the test.yaml had to be escaped by including them as `@@`. This was done to have a clear separation between literal `@` passed to the underling process (eg in a `call`) and `@` to be interpreted by sample-tester (currently, only `@args`). The idea was to allow for more substitutions later.

NOW: Since the need to pass `@` to the underlying process does come up from time to time, this PR changes the semantics so that `@foo` is passed verbatim if `@foo` is not a recognized substitution; currently only `@args` is recognized. If the user wishes to pass a recognized substitution (i.e., `"@args"`) to the underlying process, then the `@` must be escaped

tl;dr:
@args  --> substitution:             --arg1=val1 --arg2=val2
@foo   --> no substitution:          @foo
@@args --> escaped, no substitution: @args
@@foo  --> escaped, no substitution: @foo