Get configs from cli-specified file globs or files under the cwd
In the first step, we attempt to obtain manifests and testplans from
the file arguments passed in.

For either (or both) of the types for which we don't obtain files this
way, we perform a secod step: a recursive search down from the cwd to
obtain the files.

For either (or both) of the types for which we do obtain files as a
result of step 1, we do NOT perform step 2. In other words, we do not
augment the corresponding file list beyond what was obtained in step
1.

This addresses most of #52.

TODO still: When a directory is explicitly passed in, search
recursively for all the files under that directory.