Recurse only within specified paths if any of those are dirs
If any of the specified paths resolve to a directory, any recursion is
limited to those directories. Otherwise, as before, if we don't find
manifest and testplans within the specified paths, we recurse down the
cwd.

Fix a bug where we were not blogging recursively (ie `**` was not
doing what was intended)