Make it explicit in gen-manifest tests that tags are used verbatim (#125)
In particular, for manifest v3+, a user-specified tag of "env" does not produce "environment" in the samples. That only happens in v2, where we were using "env" as a shorthand for specifying the "language", which was checked against permissible values. In manifest v3 we pass any and all user-supplied tags to the manifest as-is.